% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{Cleans the text by removing special characters, html_tags, #tags, contaction words.}
\usage{
clean_text(text)
}
\arguments{
\item{text}{A text to clean.}
}
\value{
Cleaned text
}
\description{
Cleans the text by removing special characters, html_tags, #tags, contaction words.
}
