% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lyrics.R
\name{extract_lyrics}
\alias{extract_lyrics}
\title{Extracting lyrics for a song}
\usage{
extract_lyrics(token, song_title, artist)
}
\arguments{
\item{token}{Title of the song}

\item{song_title}{Artist of the song}

\item{artist}{A token from Genius for asccessing lyrics}
}
\value{
songs Return lyrics of song with three columns: song_title, artist and lyrics
}
\description{
Extracting lyrics for a song
}
