% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{Cleans the raw text into human readable text.}
\usage{
clean_text(text)
}
\arguments{
\item{text}{A text to clean.}
}
\value{
A 'text' output without special characters, html_tags, #tags, contaction words, stop-words and convert everything to lower case
}
\description{
Cleans the text by removing special characters, html_tags, #tags, contaction words, stop-words and convert everything to lower case.
}
\examples{
text <- "Early optimization is the root of all evil!"
clean_text(text)
}
