% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Downloads dataset from kaggle and creates a dataframe with input columns}
\usage{
download_data(dataset, columns)
}
\arguments{
\item{dataset}{kaggle dataset name to download}

\item{columns}{list of columns to create a dataframe}
}
\value{
A dataframe with the given column names
}
\description{
Downloads dataset from kaggle and creates a dataframe with input columns
}
\examples{
dataset <- "geomack/spotifyclassification"
df <- download_data(dataset, c("song_title", "artist"))
df
}
