% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lyrics.R
\name{extract_lyrics}
\alias{extract_lyrics}
\title{Extracting lyrics for a song}
\usage{
extract_lyrics(song_title, artist)
}
\arguments{
\item{song_title}{Title of the song}

\item{artist}{Artist of the song}
}
\value{
Return song lyrics
}
\description{
Extracting lyrics for a song
}
\examples{
extract_lyrics( "22", "Taylor Swift")
}
