% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cloud.R
\name{plot_cloud}
\alias{plot_cloud}
\title{Creates a wordcloud of most occurring words in a string or list of strings}
\usage{
plot_cloud(
  song,
  max_font_size = 1,
  max_words = 100,
  background_color = "black"
)
}
\arguments{
\item{song}{A data frame including song_title and artist in each column}

\item{max_font_size}{maximum font size - numeric}

\item{max_words}{maximum number of words to be included in wordcloud}

\item{background_color}{background color}
}
\value{
A wordcloud image
}
\description{
Creates a wordcloud of most occurring words in a string or list of strings
}
\examples{
song <- data.frame(song_title  = c("22", "Bohemian Rhapsody"), artist = c("Taylor Swift", "Queen"))
plot_cloud(song, max_font_size=1.6, max_words=100, background_color="white")
}
